Make DataBag extend GeneratedObject instead
