+++ generator/src/main/java/com/scottlogic/deg/generator/generation/databags/DataBag.java
@@ -4,44 +4,24 @@ import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.common.profile.ProfileFields;
 import com.scottlogic.deg.generator.*;
 import com.scottlogic.deg.generator.outputs.CellSource;
+import com.scottlogic.deg.generator.outputs.GeneratedObject;
 import com.scottlogic.deg.generator.outputs.RowSource;
 
 import java.util.*;
 import java.util.stream.Collectors;
 
 
-public class DataBag {
+public class DataBag extends GeneratedObject {
     public static final DataBag empty = new DataBag(new HashMap<>());
-
-    public Map<Field, DataBagValue> getFieldToValue() {
-        return fieldToValue;
-    }
-
-    private final Map<Field, DataBagValue> fieldToValue;
-
     public DataBag(Map<Field, DataBagValue> fieldToValue) {
-        this.fieldToValue = fieldToValue;
-    }
-
-    public Object getValue(Field field) {
-        if (!this.fieldToValue.containsKey(field))
-            throw new IllegalStateException("Databag has no value stored for " + field);
-
-        return this.fieldToValue.get(field).value;
-    }
-
-    public DataBagValue getValueAndFormat(Field field) {
-        if (!this.fieldToValue.containsKey(field))
-            throw new IllegalStateException("Databag has no value stored for " + field);
-
-        return this.fieldToValue.get(field);
+        super(fieldToValue);
     }
 
     public static DataBag merge(DataBag... bags) {
         Map<Field, DataBagValue> newFieldToValue = new HashMap<>();
 
         FlatMappingSpliterator.flatMap(Arrays.stream(bags)
-            .map(r -> r.fieldToValue.entrySet().stream()),
+            .map(r -> r.getFieldToValue().entrySet().stream()),
             entrySetStream -> entrySetStream)
             .forEach(entry -> {
                 if (newFieldToValue.containsKey(entry.getKey()))
@@ -51,26 +31,5 @@ public class DataBag {
             });
 
         return new DataBag(newFieldToValue);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        DataBag dataBag = (DataBag) o;
-        return Objects.equals(fieldToValue, dataBag.fieldToValue);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(fieldToValue);
-    }
-
-    public RowSource getRowSource() {
-        return new RowSource(
-            fieldToValue.entrySet().stream()
-                .map(e -> new CellSource(e.getKey() , e.getValue()))
-                .collect(Collectors.toList())
-        );
     }
 }
