+++ generator/src/main/java/com/scottlogic/deg/generator/outputs/GeneratedObject.java
@@ -1,19 +1,57 @@
 package com.scottlogic.deg.generator.outputs;
 
 import com.scottlogic.deg.common.profile.Field;
-import com.scottlogic.deg.generator.generation.databags.DataBag;
 import com.scottlogic.deg.generator.generation.databags.DataBagValue;
 
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 /** A set of values representing one complete, discrete output (eg, this could be used to make a full CSV row) */
-public class GeneratedObject extends DataBag {
+public class GeneratedObject {
+
+    private final Map<Field, DataBagValue> fieldToValue;
+
     public GeneratedObject(Map<Field, DataBagValue> fieldToValue) {
-        super(fieldToValue);
+        this.fieldToValue = fieldToValue;
     }
-    public GeneratedObject(DataBag dataBag) {
-        super(dataBag.getFieldToValue());
+
+    public Map<Field, DataBagValue> getFieldToValue() {
+        return fieldToValue;
+    }
+
+    public Object getValue(Field field) {
+        if (!this.fieldToValue.containsKey(field))
+            throw new IllegalStateException("Databag has no value stored for " + field);
+
+        return this.fieldToValue.get(field).getUnformattedValue();
+    }
+
+    public DataBagValue getValueAndFormat(Field field) {
+        if (!fieldToValue.containsKey(field))
+            throw new IllegalStateException("GeneratedObject has no value stored for " + field);
+
+        return fieldToValue.get(field);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        GeneratedObject generatedObject = (GeneratedObject) o;
+        return Objects.equals(fieldToValue, generatedObject.fieldToValue);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(fieldToValue);
+    }
+
+    public RowSource getRowSource() {
+        return new RowSource(
+            fieldToValue.entrySet().stream()
+                .map(e -> new CellSource(e.getKey() , e.getValue()))
+                .collect(Collectors.toList())
+        );
     }
 }
